\name{excelRio-package}
\alias{excelRio-package}
\alias{excelRio}
\docType{package}
\title{
Helpful Functions for Exchanging Data with Excel
}
\description{
Two functions to exchange small amounts of data 
between \R and Excel via the clipboard.
Two functions to exchange larger amounts of data 
between \R and Excel via csv files.
Has been tested on a Windows machine (R 3.0.1)
and a Mac (R 2.15).
}
\details{
\tabular{ll}{
Package: \tab excelRio\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-06\cr
License: \tab GPL (>=3)\cr
}
    readFromExcel,
    writeToExcel,
    pasteFromExcel,
    copyToExcel,
    readFromCsv,
    writeToCsv,
}
\author{
Dan Murphy

Maintainer: <danielmarkmurphy@gmail.com>

}
\keyword{ package }
\seealso{
RExcel, MRMR, Excelsi-R
}
